refactor(ba-platform): move Apollo utilities to lib/apollo/
Apollo is framework-agnostic — shouldn't be nested under lib/react/.
- apollo/react/  — web client + csrfLink
- apollo/expo/  — Expo client + expoOrgLink
- apollo/orgLink.ts, apollo/user/ — shared across platforms
- lib/react/providers/ — React-specific (unchanged)
- lib/expo/ApolloClientProvider.tsx — React component (stays in expo/)
