+++ libs/ba-platform/src/lib/apollo/expo/index.ts
@@ -1,0 +1,1 @@
+export { createApolloClient } from './client';
