+++ libs/ba-platform/src/lib/apollo/user/UserProvider.graphql
