--- libs/ba-platform/src/lib/expo/apollo/index.ts
@@ -1,1 +1,0 @@
-export { ApolloClientProvider } from './ApolloClientProvider';
