--- libs/ba-platform/src/lib/react/apollo/index.ts
@@ -1,5 +1,0 @@
-export { createApolloClient } from './client';
-export { csrfLink } from './csrf';
-export * from './orgLink';
-export * from './user/__generated__/UserProvider.generated';
-export { CurrentOrgUserDocument } from './user/__generated__/UserProvider.generated';
