+++ libs/ba-platform/src/lib/react/index.ts
@@ -1,2 +1,5 @@
-export * from './apollo';
+export * from '../apollo/react';
+export * from '../apollo/orgLink';
+export * from '../apollo/user/__generated__/UserProvider.generated';
+export { CurrentOrgUserDocument } from '../apollo/user/__generated__/UserProvider.generated';
 export * from './providers';
